Update opencore-issues.md
When installing Catalina on a Comet Lake based system, I was stuck on "failed to boostrap img4 values." The guide shows the value for Misc->Security->SecureBootMode being set to Default. By changing this to Disabled, I was able to get past the issue and install MacOS. 

Maybe a better approach is to update the install guide (in my case for Comet Lake). That's up to you.

Thank you for the detailed install guide!